Make it possible to browse the gateway
Announce the gateway as "browsable". At the moment it only lists
itself, which allows some clients (e.g. Gajim) to subscribe to it,
which they couldn't do otherwise. Another advantage is that it
allows the user to see a human description of the service.

We could announce other items in the future, for example the various
bitcoin addresses that a user has control over (those would be JIDs),
or the list of pending transactions (those would be nodes).