User's total balance and list of addresses
Implement new method "getAddresses()" that lists all addresses of a
given user.

Also fix getBalance() so that it returns the user's balance, not the
whole balance of the bitcoin installation (it was a placeholder).

At the moment the computation of getBalance() only gives the total
of all bitcoins ever received to that address. This is not entirely
correct, since it should then substract all past transactions made
by this user. However, they can't make transactions yet, so it's
equivalent.