Bitcoin package + add first address to roster
Creation of a specific package for all work that is only dealing with
bitcoins: addresses and connection to the client. This package could
be part of a separate project later.

Also, after creating a first address for the newly-registered user,
add its JID to the user's roster.

We face a big problem here: a bitcoin address is case-sensitive, the
node part of a JID isn't. We need to come up with a good encoding
function.