Cleaner handling of SQL connections
The SQL() class looks for an URL in its arguments, instead of accessing
the application's configuration. The conf reading is done by the
application itself.

Also, as a related improvement, the class can cache several connections.
Each time it's called with a given URL, it looks it up into the cache.
If found, it reuses it, otherwise it initializes a new connection and
caches it. If no URL is provided, the class uses an arbitrary existing
connection.

This module has become very generic and is not bitcoIM-related, but we
leave it here until we come up with a better home.