New class db.Database that handles db contents
Rename module 'sql' to 'db'.

Inside the module, create a new class "Database", whose aim is to
handle the contents of the database. In particular, add a function
that takes care of installing/upgrading the database.

This suppresses the need of manually creating the tables. Suppress
the SQL file and update the installation instructions.