Switch commands to lowercase
Don't care about case in the first word of a command. This is useful
to allow commands such as "HELP". Also, some jabber clients capitalize
the input by default, so it's easier if user don't need to worry about
that.

Don't change case of any other part of the command, obviously: for
instance, we don't want bitcoin addresses or free-form text to be
modified.