Labels are now called accounts
The bitcoin software replaced the notion of label with the notion of
account. Although the old version still works, stick to the new
official term.

There's no change in the way it can be used, although new API functions
might appear in the future, now that bitcoin cares about accounts.