Move command-related stuff to their own class
Introduce new class: Command. This represents an abstract command
sent by a user: action, array of arguments, user who's sending the
command, target of the command (who the user is talking to).

There is also a function in this class, that parses a command where
the command was sent as a command line. This step is optional, in
that there might be other interfaces for issuing commands, e.g.XMPP's
Ad-Hoc Commands (XEP-0050).

The two main methods are execute() and usage().