The user of a command is given to .execute()
A command is generic and does not take a user anymore. Instead, the
user is given to the .execute() method. THis way, in theory, the
same command could be called by different users. A command can also
be constructed with no user at all, e.g. is all you want is its
usage.